Remove usbreset and clamav dir.
usbreset was an attempt for usb nics to be automatically restarted, but
this binary simply didn't work. Building fos with it isn't helpful and
breaks the x86 side of things, so just simply removing is best.

clamAv directory is being removed. It's not s simple program and even
more difficult to keep updating. I've not removed the script that runs
clamav, so people could get this working on their own environments if
they'd like.