Change label test for reserved partition
It's been a long standing problem that the test fails for the label of the System Reserved partition for certain languages. 

The test has been updated to work across more languages now. It's slightly less precise, but more neat and should now work for English, French and Dutch (and possibly others that I didn't think about).

The original test had an é in the middle which never matched and thus French users had resize problems all the time. This could of course simply be edited, but I wanted to push this new regex instead as it also works for Dutch.