Make the init boot up so it doesn't wait indefinitly on network errors, thanks to @WhiteAls
If we have problems with DHCP or we can't reach the FOG web server, then the network script will wait for user input indefinitely.
In unattended scenarios (cron induced capture) this will lock system until user will notice the problem. For my case the problem is often resolved by simple reboot (seems like a spanning tree problem, but it is strange that simple reboot solves the problem).
I do undestand that user should have the ability to analyze the problem by reading on screen log, but there is no need to wait for input indefinitely. Instead, we can use `read -t 60` or less, as we will hit another 60 seconds timeout in the next script